some refactoring and cleanup
